refactor: rename secp256k1 modules
